+++ src/@core/theme/overrides/button.ts
@@ -82,21 +82,20 @@ const button: Theme['components'] = {
       }),
       sizeSmall: ({ theme }) => ({
         fontSize: '1.0625rem',
-        borderRadius: theme.spacing(4),
+        borderRadius: 'var(--mui-shape-customBorderRadius-xl)',
         minWidth: theme.spacing(26.25),
-        height: theme.spacing(10) 
-
+        height: theme.spacing(10)
       }),
       sizeMedium: ({ theme }) => ({
         fontSize: '1.0625rem',
         lineHeight: 1.529412,
-        borderRadius: theme.spacing(4),
+        borderRadius: 'var(--mui-shape-customBorderRadius-xl)',
         minWidth: theme.spacing(50),
         height: theme.spacing(10)
       }),
       sizeLarge: ({ theme }) => ({
         fontSize: '1.0625rem',
-        borderRadius: theme.spacing(4),
+        borderRadius: 'var(--mui-shape-customBorderRadius-xl)',
         width: theme.spacing(42),
         height: theme.spacing(14)
       }),
