+++ src/components/theme/types.ts
@@ -22,6 +22,7 @@ declare module '@mui/material/styles' {
         md: number
         lg: number
         xl: number
+        rounded:number
       }
     }
     customShadows: {
