+++ src/views/retailer-page/template/RetailWebsite.tsx
@@ -1,96 +1,123 @@
-import { AppBar, Toolbar, Button, Typography, TextField, Box, InputAdornment } from '@mui/material'
+import React from 'react'
 
-export default function RetailWebsite() {
+import Link from 'next/link'
+
+import { Box, Typography, InputBase, IconButton, Button, TextField, Divider } from '@mui/material'
+
+const rightSideLinks = [
+  { label: 'Find a Store', href: '/' },
+  { label: 'Track the order', href: '/' },
+  { label: 'Join Us', href: '/' },
+  { label: 'Sign In', href: '/' }
+]
+
+const navigationLabels = [
+  { label: 'Label', href: '/' },
+  { label: 'Label', href: '/' },
+  { label: 'Label', href: '/' },
+  { label: 'Label', href: '/' },
+  { label: 'Label', href: '/' },
+  { label: 'Label', href: '/' }
+]
+
+const RetailWebsite = () => {
   return (
-    <div className='min-h-screen bg-gray-100'>
-      {/* Navbar */}
-      <AppBar position='static' className='bg-white shadow-md'>
-        <Toolbar className='flex justify-between'>
-          {/* Left Section: Logo and Hotline */}
-          <Box display='flex' alignItems='center' className='space-x-6'>
-            <Box border='1px solid black' p={2}>
-              <Typography variant='h6' color='textPrimary' className='font-bold'>
-                Logo
-              </Typography>
-            </Box>
-            <Typography variant='body1' color='textSecondary'>
-              Hotline: <strong>01234567891</strong>
-            </Typography>
+    <Box
+      className='min-h-screen w-full bg-white flex justify-center items-center'
+      sx={{ height: '1024px', width: '1440px' }}
+    >
+      <Box className='w-full h-full bg-white p-4'>
+        <Box className='flex justify-between items-center mb-2'>
+          <Typography variant='body2' className='text-gray-500'>
+            Hotline: <strong>01234567891</strong>
+          </Typography>
+
+          <Box className='flex gap-4'>
+            {rightSideLinks.map((link, index) => (
+              <React.Fragment key={index}>
+                {index > 0 && <Divider orientation='vertical' flexItem className='mx' />}
+                <Link href={link.href}>
+                  <Typography
+                    variant='body2'
+                    className={`cursor-pointer ${link.label === 'Track the order' ? 'text-blue-600' : 'text-gray-500'}`}
+                  >
+                    {link.label}
+                  </Typography>
+                </Link>
+              </React.Fragment>
+            ))}
           </Box>
-
-          {/* Center Section: Navigation Links */}
-          <Box display={{ xs: 'none', md: 'flex' }} className='space-x-4'>
-            <Typography variant='body1' className='cursor-pointer'>
-              Label
-            </Typography>
-            <Typography variant='body1' className='cursor-pointer'>
-              Label
-            </Typography>
-            <Typography variant='body1' className='cursor-pointer'>
-              Label
-            </Typography>
-            <Typography variant='body1' className='cursor-pointer'>
-              Label
-            </Typography>
-            <Typography variant='body1' className='cursor-pointer'>
-              Label
-            </Typography>
-            <Typography variant='body1' className='cursor-pointer'>
-              Label
-            </Typography>
-          </Box>
-
-          {/* Right Section: Find a Store, Track Order, Join Us */}
-          <Box display='flex' alignItems='center' className='space-x-4'>
-            <Typography variant='body1' className='cursor-pointer'>
-              Find a Store
-            </Typography>
-            <Typography variant='body1' className='cursor-pointer text-blue-600'>
-              Track the order
-            </Typography>
-            <Typography variant='body1' className='cursor-pointer'>
-              Join Us
-            </Typography>
-            <Typography variant='body1' className='cursor-pointer'>
-              Sign In
-            </Typography>
-
-            <Box className='relative'>
-              <TextField
-                variant='outlined'
-                size='small'
-                placeholder='Search'
-                InputProps={{
-                  startAdornment: (
-                    <InputAdornment position='start'>
-                      <i className='ri-search-line text-gray-400'></i> {/* Using an <i> tag for the search icon */}
-                    </InputAdornment>
-                  )
-                }}
-                className='bg-gray-100 rounded-full'
-              />
-            </Box>
-          </Box>
-        </Toolbar>
-      </AppBar>
-
-      {/* Main Content */}
-      <Box className='flex flex-col items-center justify-center mt-16'>
-        <Typography variant='h4' className='font-bold'>
-          Track delivery progress of an order
-        </Typography>
-        <Typography variant='body1' className='text-gray-600 mt-2'>
-          Enter the given code to track your order
-        </Typography>
-
-        <Box className='mt-8 w-full max-w-lg'>
-          <TextField fullWidth label='Order ID*' variant='outlined' className='bg-white' />
         </Box>
 
-        <Button variant='contained' color='primary' className='mt-4 px-12 py-2 rounded-full bg-gray-300 text-gray-500'>
-          Submit
-        </Button>
+        <Box className='flex justify-between items-center mb-6'>
+          <Box
+            sx={{
+              border: '1px solid #000',
+              padding: '10px',
+              width: '180px',
+              height: '64px',
+              display: 'flex',
+              justifyContent: 'center',
+              alignItems: 'center'
+            }}
+          >
+            <Typography variant='h5'>Logo</Typography>
+          </Box>
+
+          <Box className='flex gap-4'>
+            {navigationLabels.map((navItem, index) => (
+              <Typography key={index} variant='body1' className='text-black mr-4'>
+                {navItem.label}
+              </Typography>
+            ))}
+          </Box>
+
+          <Box className='flex items-center gap-4'>
+            <Box className='flex items-center rounded-full bg-gray-100 px-4 py-2'>
+              <IconButton className='text-gray-500'>
+                <i className='ri-search-line' style={{ fontSize: '24px' }}></i>
+              </IconButton>
+              <InputBase placeholder='Search' />
+            </Box>
+
+            <IconButton>
+              <i className='ri-heart-line' style={{ fontSize: '24px' }}></i>
+            </IconButton>
+            <IconButton>
+              <i className='ri-shopping-cart-line' style={{ fontSize: '24px' }}></i>
+            </IconButton>
+          </Box>
+        </Box>
+
+        <Box className='flex flex-col items-center mt-16'>
+          <Typography variant='h2' className='text-black font-bold mb-4'>
+            Track delivery progress of an order
+          </Typography>
+
+          <Typography variant='h5' className='text-gray-500 mb-8'>
+            Enter the given code to track your order
+          </Typography>
+
+          <TextField
+            id='order-id'
+            variant='outlined'
+            placeholder='Order ID*'
+            InputProps={{
+              className: 'rounded-lg h-[60px]'
+            }}
+            className='w-[444px] h-[60px] mb-8 rounded-lg - placeholder-gray-500'
+          />
+
+          <Button
+            variant='contained'
+            className='w-[265px] h-[60px] bg-gray-300 text-white rounded-[30px] capitalize-none'
+          >
+            Submit
+          </Button>
+        </Box>
       </Box>
-    </div>
+    </Box>
   )
 }
+
+export default RetailWebsite
